Rename the Python module
PyPI didn't like the `hook_tools` name :/ Let's try with
`omegaup_hook_tools` instead.
