+++ src/omegaup_hook_tools/data/phpcbf/Standards/OmegaUp/Sniffs/Classes/ClassDeclarationSniff.php
