+++ src/omegaup_hook_tools/data/phpcbf/Standards/OmegaUp/ruleset.xml
