+++ src/omegaup_hook_tools/data/tidy.txt
